This adds Dynamic Loading of the library DLL's.
If the libraries are statically loaded, this will change nothing. (So we can change over projects one at a time). The Project AddArt is modified for delay loading.

It allows, but does not require, for a path to be provided to override the default DLL location (..\..\Libs). So no sample need to modified to use this.

In order to not interact with plugin loading, the plugin path supplied to the PDFLDataRec is set as a full, rather than a relative, path name.

The HINSTANCE returned from LoadLibrary is forwared to APDFL via the inst filed of the PDFLDataRec.

And misc., the destructure for the APDFLLib Object will now call PDFLTermHFT ONLY is the Init succeeded (Avioding a crash if the init failed).

Signed-off-by: Michael McKeough <mjm@datalogics.com>