Quick cleanup of some ugly make constructs.
Makefile is the preferred name for a makefile, because it is the 2nd file that make will look for.  This allows a user to copy Makefile to makefile and make changes for experimentation.
Modern make implementations set some varialbes implicitly, and have implicit rules defined.  It's familiar to see these conventions used and helpful when poting to oddball platforms that require different commands.  Mac OSX for example does not have "gmake" installed by default, and even many GNU make installations from vendors are not installed as "gmake".  Without notable exceptions, the variable $(MAKE) can be used inside a makefile, and is set to whatever "make" ran the top-level command.  It's a good rule to always use $(MAKE) to run make from a makefile.
Also removed the Linux-specific libpath setting that wasn't in the linux-specific makefile, and found that it wasn't necessary anyway, as I stated in an earlier PR.