Update usage of Verdana to CourierStd and small fixups.
Original problem was Verdana (TrueType) was not available on all platforms so Courier was used instead. Unfortunately the output created by using Courier wasn't observed and the text still referred to using Verdana which is confusing to the user and went beyond the page bounds. The problem is addressing by removing lingering references to Verdana and by splitting it into two text runs that fit comfortably on the page.

Use calculated values for gStateLen and tStateLen instead of 0 and initialize structs before using them.