Update program description and comments for AddThumbnailsToPDF sample program
Wrote proposed new description of this sample and reviewed comments as part of Repackaging project.

Lines 7 to 12, the initial comment, This function is only a sample and not supported
would apply to all of our sample program code. All samples are intended to provide a
model for working but are not meant to be used in production. I plan to add a statement
to this effect to the Developers page where we describe all of the APDFL sample C++ programs.

The initial comment about reading a book on image processing or a programming reference
seems unnecessary. The statement is too vague to be useful. None of the other sample programs
offer similar statements.

This is an initial pull request for the 21 sample programs, and something of a test.

More pull requests to follow.