Update program description and comments for AddThumbnailsToPDF sample
Wrote proposed new description of this sample and reviewed comments as part of repackaging project.

Lines 7 to 12, initial comment about the sample not being supported would apply to all samples.
Samples are intended to serve as a model but are not meant to be used in production systems.
I plan to add a general statement to this effect on dev.datalogics.com where we discuss the
APDFL C++ sample programs.

Initial comment about reading a book on image processing or a C programming reference is too
vague to be useful. None of our other samples offer similar statements.

This is an initial pull request for the 21 sample programs.