[Packaging] Update program description and comments for CountColorsinDoc and CreateBookmarks samples
Wrote proposed new descriptions of these two samples and reviewed comments as part of repackaging project.

For CountColorsinDoc, consider edits to lines 165-170.

For CreateBookmarks, JIRA card APDFL-1197 notes that this sample program may be redundant. We could replace this sample with AddBookmarks.

This pull request could still be merged, and then the decision regarding AddBookmarks and CreateBookmarks made later.