[Packaging] Update program description and comments for three sample programs
EmbedFonts.cpp
ExtractFonts.cpp
DisplayPDEContent.cpp

Wrote proposed new descriptions for these samples and reviewed comments as part of repackaging project.

For DisplayPDEContent added a description of bounding regions.

For EmbedFonts, a more detailed description will be provided on the Developer's page. I will provide a link to that page in the program description in the sample in a later pull request.

For EmbedFonts, I deleted a TODO statement at line 12. We could create a task for this if needed.