[Packaging] Update program description and comments for helloJapan sample
Wrote proposed new description of this sample and reviewed comments as part of repackaging project.

Note question on line 110 of original program. I removed the question, but we should confirm that KozMinPr6N-Regular is a CIDType2 font, a Glyph ID-based font.

Also, when I publish a description of this sample to the Developer's page (dev.datalogics.com), I plan to add some more detailed content related to some of the font concepts addressed in this sample. In a later pull request I will add a note to the beginning of the program comments to provide the web page address for this content.