[Packaging] Update program description and comments for ImportPages sample
Wrote proposed new description of this sample and reviewed comments as part of repackaging project.

When I publish a description of this sample to the Developer's page (dev.datalogics.com) I will add more detailed content related to bounding boxes. In a later pull request I will add a note to the program comments, after the program description, to provide the web page address for this content.