[Packaging] Update program description and comments for PDFMakeOCGVisible sample
Wrote proposed new description of this sample and reviewed comments as part of repackaging project.

Note comment on line 95, "fluke, but handle gracefully." Not sure how to interpret it.
If "handle gracefully" is meant to guide a developer's decisions, we should be more specific.
If in fact the statement is needed at all.