[Packaging] Update program description and comments for ValidateSignatures sample
Wrote proposed new description of this sample and reviewed comments as part of repackaging project.

Consider TODO statement at line 210:

    // TODO: Here, the stored digest would NEED to be decrypted before checking for correctness,
    //       That task, however, is out of scope of this sample, and so this function will
    //       always be returning 'false'.

See JIRA card APDFL-1217. Besides removing the TODO reference, I'm not clear on how the program could work at all if decrypting the stored digest is "out of the scope of this sample."