[Packaging] Edit comments and description, PDFViewer.cpp
This program had an inordinately long description at the start of the code, in comments.

I removed most of this content to revise and publish on the Developer's page.

Added a copyright statement, and a link to the program description on dev.datalogics.com.

Read through and edited the comments, found a few errors and minor problems.