Correct some serious problems in APDFLDoc class.
Valgrind reported source & destination overlap for wscpy in APDFLDoc::setASPathName()
The member had already been wscpy'd in the previous statement.  While I had my tools out, I saw use of explicit this pointer, and deduced that it was required because a member had the same identifier as a dummy argument, so the dummy argument was changed and this-> removed.  Also there were two copies of APDFLDoc.cpp and a few other files that did not seem to be used, so I removed those as well.