Redoing adding of PDFPort directly on Mac, xcode edits and all.
What caused this sample to crash on Mac before was the missing PDFPort, so the sample was updated to reference it relative to its installed path.  But when I open the sample delivered in the release it still crashes and when I look at the settings in xcode there's blank reference where the PDFPort should be.  If I manually remove that blank reference, and then re-add the PDFPort reference it works.  I'm not sure what went wrong the first time, so I'm just checking in all of the edits xcode decided to make to the project and suggest we don't edit them and just merge it to hopefully make this IDE happy.