A more plugin like fix for the crashing plugin samples.
The problem running these two samples is that they load a plugin, which is found without issue, but then that plugin loads libraries that aren't in the search path.  This can be fixed by setting the DYLD_FRAMEWORK_PATH environment variable, or by copying the dependent libraries PDFPort and PSFSettings into the application's Framekworks directory (which is what it previously did by referencing them).  Those two solutions work, but are not very clean, so the best way to demonstrate this feature is to add the right relative path into the sample so that it can look up libraries that only the plugin cares about.  This preserves the disconnection between the plugin and the app that uses it, which is what plugins are all about.