SF40446 -- Add a new sample "OutputPreview"
This sample demonstrated using PDPageDrawContentToMemoryWithParams() using a DeviceN color space name to create a deviceN image of a page, which can then be rendered so as to show the page in "true colors", as an output preview. It demonstrates how to turn selected colorants on and off in the display of the DeviceN bitmap. it also shows separating that bitmap into single color maps, and displaying those maps using Separation color spaces, so as to maintain their "true color".

Signed-off-by: Michael McKeough mjm@datalogics.com