Create the SysEncoding from the Base Name, not the CMap Name.
Michael recently changed this sample and I think he didn't meant to use PDSysEncodingCreateFromCMapName() because he passes it PDEFontAttrs.encoding, which can be WinAnsiEncoding, MacRomanEncoding which are font encoding names.  But the function expects CMap names.  I noticed it causing a failure on Mac CI but then when I looked closer the sample also fails to run out of the box on Windows too.  Change it to use PDSysEncodingCreateFromBaseName() instead which does actually expect a font encoding name.