Fix which value we retrieve the string from.
Support noticed for a particular document, this sample crashed.  I noticed we seem to have mixed up the Key vs the Value for retrieving the ASText in our callback.  Also, for the Name case we were destroying the Key instead of the Value.  I think was perhaps not a problem with other documents because of the presence of Marked Content dictionary and we try to enumerate the Cos objects found in it.