SF41621 -- Rendered pages are white, or application fails?
The pages in question are quite large. We did not have sufficient memory to render at the requested resolutions in a 32 bit environment. In two cases, this resulted in a null pointer to a buffer, when resulted in a white image. In the third case, the amount of memory needed exceeded malloc's internal limit, and resulted in an interrupt.

This sample code is woeful. It uses ASFixed interfaces only, which with these pages could result in ASFixed overflow ox rectangles. It fails to notice if the buffer is not allocated. It uses PDPageDrawContentToMemory().

I Modified the sample to do all of it's math in floating point and use PDPageDrawContentsToMemoryWithParams. There were a number of private methods, called once, which consisted of only a few lines. I removed these. I increased the level of commentary as to what was being done, and why, and where additional information could be found. Since we do not yet have a complete set of Floating Point interfaces to APDFL, I included a set of conversion routines to convert between ASFixed, ASDouble, and ASReal rectangles and matrices (as needed).

I beleive this is now a much stronger sample for "best practice" in rendering a page. It properly captures, and reports the issues caused by the inability to allocate a bitmap buffer for the case at hand, and works properly for a large set of pages at hand.

Signed-off-by: Michael McKeough <mjm@datalogics.com>