Update to add callback support for setting the position for files over 2GB.
The sample causes PDFL to raise an error opening the file, because we try to look from the end of the file for the trailer, but the setpos callback only supports files under 2GB so it forces the value to 0.  So instead we look for the end of the file marker starting from the beginning of the file and miss our chance to find it.  This isn't really a problem because the setpos callback tells the user it simply doesn't support 2GB files.

But we should probably update the sample to at least include the newer 64-bit pos callback that does support files over 2GB so it doesn't fail out of the box.