Update common method to use PDSaveFlags2 when saving, use PDSaveCompressed in MergeDocuments sample for best results.
Customer noticed the output of merging documents could be unexpectedly large but if they switched the source and destination it would be the expected size.  In this case one document uses objects streams and the other doesn't.

What's happening is when we save using our common method we don't use PDSaveFlags2 so we look at if our document has full compression when we decide to compress the objects or not.

When our destination document (from mering) is compressed we apply the compressed flag when saving.

When our destination is not compressed we don't apply the compress flag when saving.

By changing it to specify compression during saving we'll attempt to use it for either scenario to get a more consistent result.

Update MergeDocuments to use PDSaveCompressed flag.

Update common sample method to accept PDSaveFlags2.