Correct Hello Japan Sample
I came across this looking for another issue in the sample (A customer would like to know why we chose those CMaps, and how he can choose others). We write two pages here. the start of the second page reads:

    // The second page of the document uses a different font,
    // KozMinPr6N-Regular. This is a CIDType2 font, or a glyph ID-based font.
    pdPage = PDDocCreatePage(pdDoc, PDBeforeFirstPage, mediaBox);

Which, of course, creates the second page BEFORE the first page. So the description and the sample do not match.

This change corrects it to read:

      pdPage = PDDocCreatePage(pdDoc, PDDocGetNumPages(pdDoc)-1, mediaBox);

Which will put the second page second.

Signed-off-by: Michael McKeough <mjm@datalogics.com>