Work to improve samples for APDFL v18
In APDFL 18, the OS variable is set with _64 appeneded for 64-bit unix.
This causes problems if we don't strip it off, when we need to.
Simplify what sample gets built where, and sort the resulting list of samples.
Add the XPS2PDF sample for Linux & Solaris, currently crahses under
Linux 64 in Release builds - but otherwise this makes APDFL_Samples dev test pass there.