AddDirectWatermark
Add a new sample program "AddDirectWatermark".

This sample demonstrates adding a text string as a watermark, directly in the content of each page, rather than using the Watermark Annotation. Such a watermark is a great deal harder to find and remove from a document. The sample also saves the document with encryption. Permitting only display and printing of the document.

Signed-off-by: Michael McKeough <mjm@datalogics.com>