Replace the input document for FlattenAnnotations.
The structure and content of the document used in this sample, not due to the needs of the sample itself, causes the sample to take nearly 5 minutes to save it's output. After about 5 seconds of actual processing.

This is annoying in testing, as it pushed APDFL-Samples close to it's tie limit. But it is also a bad thing to set before customers. (Oh. Look. It takes 5 minutes to flatten the annotations in a a one page document!).

The reason it take so long is that the document has a large Structure tree, and Name tree. The page was extracted from a much larger document, and the entire tree is still present in it.

This change removes those elements, with no change to the appearance of behavior of the sample. Except than now it completes in 7 seconds, rather than 5 minutes.

Signed-off-by: Michael McKeough <mjm@datalogics.com>