Don't resolve invalid structure parent references.
Customer found splitting a document took a very long time and actualy ran out of space in temp waiting for it to finish.

As a best practice don't resolve invalid structure parent references, this is quite slow to do and for very complex structure trees it leads to explosively large temp files created.  For many documents structure doesn't matter at all for a merge and for the customer's document it means processing completes in seconds.