Add support for building the samples with APDFL from Conan
We're not creating a package with Conan, we just want to USE a package,
so we add conanfile.txt and put the generators and dependencies in.
Also need to figure out how to get the "source" that APDFL ships.
Use the host command line argument.
Add the import statements for plugins binaries.
32-bit solution gets copied, so delete it explicitly.