Set PDSaveAddFlate by default.
Customer presented a document with ~100MB uncompressed streams, by saving with PDSaveAddFlate we yield a big savings.  The customer was creating a new document so they realized the savings and if they used Optimizer they'd also realize it, but in general this is a good practice for all of our samples so let's just set it to default to use this.