Remove ValidateSignatures sample.
The addition of this sample appears to have been short-sighted.  It doesn't quite verify and even has a comment in the code that says it will always tell you the Signature is invalid and invites the user to finish the work.

I spoke to Patrick and Scott and have come to the conlucions this should _not_ be a shipped sample and should be removed.

Remove sample file used with it.