Fix how PageInfo struct members are initialized.
Trying to investigate SF44177 using the sample I noticed it crashed because we memset() the members of the PageInfo which is problematic because some of the members are vectors so pushing content onto the vector goes wrong because its members were reset.

Remove the memsetting and add a constructor to initialize the members.