Remove WebOptimizedPDF sample and update PDFOptimizer sample.
1) The WebOptimizedPDF sample was already trivial, to save a linearized pdf you simply use the PDSaveLinearized flag, so we have a whole sample to show how to save a PDF as you normally would and use this flag.

Having this sample is confusing because using the Optimizer API you can already specify to save as Linearized, also it sets up the confusing question of which type of optimization do I need? If my documents are going on the web, you could draw the conclusion the Optimizer sample is not appropriate.

Now there will be one sample in the DocumentOptimization category, PDFOptimizer.

2) Update the PDFOptimizer sample copyright and description, clarify comments to be more accurate.