Jenkinsfile: Prevent mkenv locations from mixing between jobs
This fixes a problem where a 32-bit Solaris job ran on Tibet, and a 64-bit
Solaris job ran on Lhasa. They tried to share the same slot in ENV_LOC, and then
one job tried to activate a Python environment with the incorrect name.

Since the matrix has two axes, NODE and BITS, make the key for ENV_LOC out of a
combination of NODE and BITS to prevent the two jobs from sharing the same key
in the dictionary, conflicting over what the Python environment directory is.