Fix reporting of color value.
QA reported a regression of how an indexed color value was expressed between releases.

The PDEColorValue.color element is Fixed so by simply grabbing its value and multiplying by the number of components we're going to end up with a big number and try to access our ColorTable at that index.  So we're just pointing at garbage in memory.

To fix this let's convert from fixed and then use that to access the right colortable index.