Add new ConvertToFactur-X sample.
Factur-X is a new Franco-German standard which is also known as Zugferd v2.2 but is different enough from it to have a new name.  Customer pointed out our existing Zugferd sample is out of date and we should support this.  Its sample is similar to the Zugferd one, we pretty much rely on PDF/A-3 conversion and we just need to tweak a few things to get it to be valid.

The input XML document is required to be specified by the user so similar to the to the other sample we won't run this sample automatically.

This sample differs from its older relative a bit in that there's an optional argument for specifying the relationship type which is much more important now.  Also the input XML file name is specified by the standard as well.