Add new HeaderFooter sample.
Sample illustrates how to create Header and Footer text. This is a common workflow that other competitors offer examples or samples of how to do it. Its not the easiest thing to do either for somebody new.

To facilitate this add a new method to the Font class, MeasureTextWidth() to determine how wide the text measures at the specified font size for proper placement of the header and footer.

The usual changes for a new sample: Update VS solution files which point to sample projects, update make file, and all scripts accordinly.