Fix a couple problems.
As part of creating the DeviceN image we were always freeing the buffer if it wasn't null, just set it to null after freeing
it in the loop to avoid a double free.

Because Spot Colorants may only be part of Annotation appearances, to do a proper DeviceN rendering of all colorants this needs to be set.
Otherwise we collect all the spot colorants including those of Annotation appearances and setup our DeviceN rendering but at the same time
we tell PDFL with our render flags "don't consider them", leading to scrambled output.