Revamp sample and fix a subtle problem.
1) Investigating a customer issue, I noticed my local output was completely wrong and almost all of its glyphs incorrect.

What's not so obvious is out default input uses one type of Font technology as a basis for its FontDescendant, but what happens if we discover an OS font that's a match but uses different
technology?  Although PDFL can make the appropriate changes to the Font itself, that doesn't change what the characters codes are in the actual Page Content.

Add some checks for this situation and avoid embedding in this case.

2) This sample was poorly written, it made it look much more complicated than it needed to be.  Greatly simplify it so it doesn't look more complicated than it should be.