Account for appearance BBox.
Customer found Flattening an annotation appearance resulted in a misplaced output.

Normally we use the annotation's rectangle to set the postional offset of our concatenate matrix, but in this case the BBox of the appearance form is a bit offset from its own 'origin'.

So we end up duplicating the positional offset and it gets thrown a good ways away from the expected location.

Check for this scenario and just force the BBox to start at 0.